Rename "skimmer" to "newswaters-job".
