+++ newswaters-job/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "skimmer"
-version = "0.6.3"
+name = "newswaters-job"
+version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
