+++ newswaters-job/Dockerfile.deployment
@@ -1,9 +1,9 @@
 # Build
 FROM rust:1.72.0 AS build
 
-COPY . /usr/src/skimmer/
+COPY . /usr/src/newswaters-job/
 
-WORKDIR /usr/src/skimmer/
+WORKDIR /usr/src/newswaters-job/
 
 RUN cargo install diesel_cli --no-default-features --features postgres
 RUN cargo build --release
@@ -15,7 +15,7 @@ RUN apt update -y && \
     apt install -y libpq5 ca-certificates chromium
 
 COPY --from=build /usr/local/cargo/bin/diesel /bin/diesel
-COPY --from=build /usr/src/skimmer/target/release/skimmer /bin/skimmer
-COPY ./migrations /usr/src/skimmer/migrations
+COPY --from=build /usr/src/newswaters-job/target/release/newswaters-job /bin/newswaters-job
+COPY ./migrations /usr/src/newswaters-job/migrations
 
-WORKDIR /usr/src/skimmer/
+WORKDIR /usr/src/newswaters-job/
