+++ newswaters-job/src/repository/item.rs
