+++ newswaters-job/src/schema.rs
