+++ newswaters-job/src/service/hacker_news.rs
